requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python-headless
tensorflow-cpu==2.12
python-multipart
